added authentication and RBAC

+++ reviews_service/reviews_app.py
@@ -5,23 +5,14 @@ import datetime
 from db_config import db, init_db
 from customer_service.customer_app import User
 from inventory_service.inventory_app import Product
+from sqlalchemy.exc import IntegrityError
+from flask_jwt_extended import JWTManager, create_access_token, jwt_required, get_jwt_identity
 
 
 app = Flask(__name__)
 init_db(app)
 app.config['SECRET_KEY'] = "222222222233333333"
-
-class User(db.Model):
-    __tablename__ = 'User'
-    full_name = db.Column(db.String(100), nullable=False)
-    username = db.Column(db.String(50), primary_key=True, nullable=False)
-    password = db.Column(db.String(100), nullable=False) 
-    isadmin = db.Column(db.Boolean, default = 0)
-    age = db.Column(db.Integer, nullable=False)
-    address = db.Column(db.String(200), nullable=False)
-    gender = db.Column(db.String(10), nullable=False)
-    marital_status = db.Column(db.String(10), nullable=False)
-    wallet_balance = db.Column(db.Float, default=0.0)
+jwt = JWTManager(app)
 
 
 class Review(db.Model):
@@ -46,14 +37,41 @@ class Review(db.Model):
 
 
 
+def admin_required(fn):
+    @jwt_required()
+    def wrapper(*args, **kwargs):
+        username = get_jwt_identity()
+        user = User.query.filter_by(username=username).first()
+        if not user or not user.isadmin:
+            return jsonify({"error": "Admins only!"}), 403
+        return fn(*args, **kwargs)
+    return wrapper
 
-# only authenticated customers
+
+@app.route('/login', methods=['POST'])
+def login():
+    data = request.json
+    username = data.get('username')
+    password = data.get('password')
+
+    user = User.query.filter_by(username=username).first()
+    inputed_password_hash= generate_password_hash(password, method='bcrypt') 
+    if not user or (inputed_password_hash!= user.password):
+        return jsonify({"error": "Invalid username or password"}), 401
+
+    token = create_access_token(identity=username, additional_claims={"isadmin": user.isadmin})
+    return jsonify({"message": "Login successful", "token": token})
+
+
+
 @app.route('/reviews', methods=['POST'])
+@jwt_required()
 def submit_review():
+    username = get_jwt_identity()
     try:
         data = request.json
         review = Review(
-            username=data['username'],
+            username=username,
             product_name=data['product_name'],
             rating=data['rating'],
             comment=data['comment']
@@ -66,25 +84,34 @@ def submit_review():
     except IntegrityError:
         return jsonify({"error": "Database error"}), 500
 
-# authenticated customers
+
 @app.route('/reviews/<int:review_id>', methods=['PUT'])
+@jwt_required()
 def update_review(review_id):
+    username = get_jwt_identity()
     review = Review.query.get_or_404(review_id)
+    if review.username != username:
+        return jsonify({"error": "Unauthorized to update this review"}), 403
+
     data = request.json
     review.rating = data.get('rating', review.rating)
     review.comment = data.get('comment', review.comment)
     db.session.commit()
     return jsonify({"message": "Review updated successfully", "review": review.as_dict()})
 
-#authenticated customers
+
 @app.route('/reviews/<int:review_id>', methods=['DELETE'])
+@jwt_required()
 def delete_review(review_id):
+    username = get_jwt_identity()
     review = Review.query.get_or_404(review_id)
+    if review.username != username:
+        return jsonify({"error": "Unauthorized to delete this review"}), 403
     db.session.delete(review)
     db.session.commit()
     return jsonify({"message": "Review deleted successfully"})
 
-#anyone 
+
 @app.route('/reviews/product/<string:product_name>', methods=['GET'])
 def get_product_reviews(product_name):
     reviews = Review.query.filter_by(product_name=product_name, is_flagged= False).all()
@@ -92,18 +119,21 @@ def get_product_reviews(product_name):
 
 #admin
 @app.route('/reviews/customer/<string:username>', methods=['GET'])
+@admin_required
 def get_customer_reviews(username):
     reviews = Review.query.filter_by(username=username).all()
     return jsonify([review.as_dict() for review in reviews])
 
-#admin
+
 @app.route('/reviews/<int:review_id>', methods=['GET'])
+@admin_required
 def get_review_details(review_id):
     review = Review.query.get_or_404(review_id)
     return jsonify(review.as_dict())
 
 
 @app.route('/reviews/flag/<int:review_id>', methods=['PUT'])
+@admin_required
 def flag_review(review_id):
     review = Review.query.get_or_404(review_id)
     review.is_flagged = True
